fixed build and eslint
