+++ packages/dom/package.json
@@ -2,6 +2,7 @@
   "name": "@utils/dom",
   "version": "1.0.0",
   "main": "dist/index.js",
+  "license": "MIT",
   "scripts": {
     "build": "tsc"
   },
