+++ packages/dom/src/checkContenteditableSliceForEmptiness.ts
@@ -1,0 +1,49 @@
+import { isCollapsedWhitespaces } from './isCollapsedWhitespaces';
+
+/**
+ * Checks content at left or right of the passed node for emptiness.
+ * @param contenteditable - The contenteditable element containing the nodes.
+ * @param fromNode - The starting node to check from.
+ * @param offsetInsideNode - The offset inside the starting node.
+ * @param direction - The direction to check ('left' or 'right').
+ * @returns true if adjacent content is empty, false otherwise.
+ */
+export function checkContenteditableSliceForEmptiness(contenteditable: HTMLElement, fromNode: Node, offsetInsideNode: number, direction: 'left' | 'right'): boolean {
+  const range = document.createRange();
+
+  /**
+   * In case of "left":
+   * Set range from the start of the contenteditable to the passed offset
+   */
+  if (direction === 'left') {
+    range.setStart(contenteditable, 0);
+    range.setEnd(fromNode, offsetInsideNode);
+
+  /**
+   * In case of "right":
+   * Set range from the passed offset to the end of the contenteditable
+   */
+  } else {
+    range.setStart(fromNode, offsetInsideNode);
+    range.setEnd(contenteditable, contenteditable.childNodes.length);
+  }
+
+  /**
+   * Clone the range's content and check its text content
+   */
+  const clonedContent = range.cloneContents();
+  const tempDiv = document.createElement('div');
+
+  tempDiv.appendChild(clonedContent);
+
+  const textContent = tempDiv.textContent ?? '';
+
+  /**
+   * In HTML there are two types of whitespaces:
+   * - visible (&nbsp;)
+   * - invisible (trailing spaces, tabs, etc.)
+   *
+   * If text contains only invisible whitespaces, it is considered to be empty
+   */
+  return isCollapsedWhitespaces(textContent);
+}
